support empty interfaces tagged as block
This commit is just to provide some extra details
about the previous commit:

We think that this check was unnecessary, and accidentally catching
fields that were typed as empty interfaces (or slices of empty
interfaces). Removing this check alleviated our problem with the encoder
not properly traversing fields that were specified as 'block' and also
typed as empty interface.